codebook implementation
